# Optimized encoding & decoding goroutines number